+++ changelog.md
@@ -5,6 +5,9 @@
   * Added instances for `Ptr` and `FunPtr`
     ([#10](https://github.com/haskell/deepseq/pull/10))
 
+  * Added instance for `IORef` and `MVar`
+    ([#6](https://github.com/haskell/deepseq/issues/6))
+
 ## 1.4.1.2  *Aug 2015*
 
   * Avoid the broken combination of GHC-7.2 with `array>=0.4`